Move classes and clean imports
